requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor==6.7.2
django-decouple==2.1
django-environ==0.11.2
django-js-asset==3.0.1
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
Pygments==2.18.0
sqlparse==0.5.1
typing_extensions==4.11.0
tzdata==2024.1
